+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## master
+
+- Allow passing in domain via constructor
+
 ## 1.1.0
 
 - Added Fax support